test: fix async tests
I must have misunderstood a lint message, so the no-floating-promises rule was disabled. Meanwhile,
I did not await the async tests, so Jest marked them as passing even if they failed. Removing the
override in the linter causes it to recognise the missing await keywords, letting Jest show the
correct result.

(Thankfully, the tests were passing.)